files added

+++ src/main.java
@@ -1,9 +1,9 @@
 
 public class main {
     public static void main(String[] args) {
-        long balance = 2_000_000_000;
-         long transfer = 500_000_000;
-                long total = balance + transfer;
+        int balance = 2_000_000_000;
+         int transfer = 500_000_000;
+                int total = balance + transfer;
         System.out.println(total);
 
  }
